Use `prepare`, not `postinstall` lifecycle script
A `postinstall` script will run even if the package is installed as a
dependency. This is undesirable because the `yarn sdks` command is only
required while developing the package, and will in fact fail when
invoked from a depending projects.

Replace by a `prepare` script, which is guaranteed to be ignored if this
package is installed as a dependency.